+++ AmeisenBotX.Core/StateMachine/States/StateBattleground.cs
@@ -15,7 +15,7 @@
             if (WowInterface.XMemory.Read(WowInterface.OffsetList.BattlegroundStatus, out int bgStatus)
                 && bgStatus == 0)
             {
-                StateMachine.SetState(BotState.Idle);
+                StateMachine.SetState((int)BotState.Idle);
                 return;
             }
